Move string enum class to utils module + add test
