+++ airflow/providers/amazon/aws/hooks/ecs.py
@@ -20,7 +20,6 @@ from __future__ import annotations
 import time
 from collections import deque
 from datetime import datetime, timedelta
-from enum import Enum
 from logging import Logger
 from threading import Event, Thread
 from typing import Generator
@@ -31,6 +30,7 @@ from botocore.waiter import Waiter
 from airflow.providers.amazon.aws.exceptions import EcsOperatorError, EcsTaskFailToStart
 from airflow.providers.amazon.aws.hooks.base_aws import AwsGenericHook
 from airflow.providers.amazon.aws.hooks.logs import AwsLogsHook
+from airflow.providers.amazon.aws.utils import _StringCompareEnum
 from airflow.typing_compat import Protocol, runtime_checkable
 
 
@@ -53,23 +53,6 @@ def should_retry_eni(exception: Exception):
             for eni_reason in ["network interface provisioning", "ResourceInitializationError"]
         )
     return False
-
-
-class _StringCompareEnum(Enum):
-    """
-    Enum which can be compared with regular `str` and subclasses.
-
-    This class avoids multiple inheritance such as AwesomeEnum(str, Enum)
-    which does not work well with templated_fields and Jinja templates.
-    """
-
-    def __eq__(self, other):
-        if isinstance(other, str):
-            return self.value == other
-        return super().__eq__(other)
-
-    def __hash__(self):
-        return super().__hash__()  # Need to set because we redefine __eq__
 
 
 class EcsClusterStates(_StringCompareEnum):
